==> Code/Part4_visualization.ipynb <==
# Part 4: Data Visualization

import pandas as pd
import matplotlib.pyplot as plt

try:
    df = pd.read_csv('cleaned_items.csv')
except FileNotFoundError:
    df = pd.read_csv('20191226-items.csv')

df.head()

## Top 10 Brands by Product Count

brand_counts = df['brand'].value_counts().head(10)
brand_counts.plot(kind='bar', figsize=(8,5))
plt.title('Top 10 Brands by Product Count')
plt.xlabel('Brand')
plt.ylabel('Count')
plt.xticks(rotation=45)
plt.show()

## Price Distribution

plt.figure(figsize=(8,5))
plt.hist(df['price'], bins=40, edgecolor='black')
plt.title('Price Distribution')
plt.xlabel('Price')
plt.ylabel('Frequency')
plt.show()

## Boxplot: Price Variation Across Top 5 Brands

top5_brands = df['brand'].value_counts().head(5).index
df[df['brand'].isin(top5_brands)].boxplot(column='price', by='brand', figsize=(8,6))
plt.title('Price Variation Across Top 5 Brands')
plt.suptitle('')
plt.xlabel('Brand')
plt.ylabel('Price')
plt.show()
